extract sync
