+++ sync/Makefile
@@ -1,0 +1,6 @@
+all: templates
+
+templates:
+	env GOOS=linux GOARCH=amd64 go build ../changes/cmd/changed
+	go-bindata -nometadata -pkg sync -o templates.go changed
+	rm changed
